bt-agent: Fix incorrect Unix signal handling
The signal handler cannot do any sort of memory allocation, and it's
likely that it refuses to stop when wrapped in a systemd service
because of this.

Instead, use the g_unix_signal_add() helper which will call the handler
in a context that does allow us to re-read the PIN text file for
example.

Note that monitoring SIGUSR1 requires glib 2.36 or newer, as mentioned
in the documentation, even if g_unix_signal_add() is already available
in glib 2.30.